refactor: split OpenCodex spend source
